+++ src/kernel/uapi_syscall.h
@@ -34,7 +34,10 @@ enum {
     SYS_PUTC      = 4,
     SYS_READ      = 5,
     SYS_YIELD     = 6,
-    SYS_EXEC      = 7
+    SYS_EXEC      = 7,
+    SYS_OPEN      = 8,   /* open a FAT16 file; returns fd or -1 */
+    SYS_CLOSE     = 9,   /* close an fd */
+    SYS_FREAD     = 10   /* read bytes from an open fd into a user buffer */
 };
 
 #endif